% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPoints.R
\name{plotPoints}
\alias{plotPoints}
\title{plotPoints}
\usage{
plotPoints(
  occ,
  location = "location",
  lon = "lon",
  lat = "lat",
  colours = "bicha"
)
}
\arguments{
\item{occ}{table containing columns with the species name, longitude, and
latitude}

\item{location}{character, name of the column containing location IDs,
default is "location".}

\item{lon}{character, name of the longitude column, default is "lon".}

\item{lat}{character, name of the latitude column, default is "lat".}

\item{colours}{character, options are "bicha" or "standard". Default is
"bicha"}
}
\value{
This function downloads all 19 climatic variables and returns them
in a raster stack.
}
\description{
Plot occurrence points with background for visualisation
}
\examples{
worldclim <- getWorldClim(res = 0.5,country = "Czech Republic",
path = getwd())
}
