% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractValues.R
\name{extractValues}
\alias{extractValues}
\title{extractValues}
\usage{
extractValues(path, occ, species = "species", lon = "lon", lat = "lat")
}
\arguments{
\item{path}{character, the path to the folder containing the variables}

\item{occ}{table containing columns with the species name, longitude, and
latitude}

\item{lon}{character, name of the longitude column, default is "lon".}

\item{lat}{character, name of the latitude column, default is "lat".}

\item{specie}{character, name of the column containing species names,
default is "species".}
}
\value{
This function returns the values for variables in each point.
}
\description{
Extracts bioclimatic data from WorldClim layers for data points.
}
